+++ implementation/config.py
@@ -1,89 +1,9 @@
 from dataclasses import dataclass
 import os
 from pathlib import Path
-import re
 from tomlkit import comment, document, nl, item, dump, load
 import pymupdf
-
-PT_BY_INCH = 72
-PT_BY_CM = PT_BY_INCH / 2.54
-PT_BY_MM = PT_BY_INCH / 25.4
-
-
-def get_value_unit(s: str):
-    value_match = re.search(r"[\d\.,]+", s)
-    if not value_match:
-        raise ValueError(f"Could not parse dimension {s}")
-    value = float(value_match.group(0).replace(",", "."))
-    just_unit = s.replace(value_match.group(0), " ").strip()
-    return (value, just_unit)
-
-
-class Dimension:
-    def __init__(self, horizontal: float, vertical: float | None, unit: str) -> None:
-        unit = unit.casefold().strip()
-        if unit == "mm":
-            unit_mult = PT_BY_MM
-        elif unit == "cm":
-            unit_mult = PT_BY_CM
-        elif unit == "inch" or unit == '"':
-            unit_mult = PT_BY_INCH
-        elif unit == "pt" or unit == "":
-            unit_mult = 1
-        else:
-            raise ValueError(f"Invalid unit type {unit}")
-        self.horizontal = horizontal * unit_mult
-        self.vertical = (vertical or horizontal) * unit_mult
-
-    @classmethod
-    def from_tuples(cls, horizontal: tuple[float, str], vertical: tuple[float, str] | None):
-        part_tuples = [horizontal, vertical] if vertical else [horizontal]
-        if len(part_tuples) == 1:
-            actual_value = part_tuples[0]
-            dim = Dimension(actual_value[0], actual_value[0], actual_value[1])
-        else:
-            actual_value_x = part_tuples[0]
-            actual_value_y = part_tuples[1]
-            if actual_value_x[1] != actual_value_y[1]:
-                raise ValueError(f"Cannot mix unit in Margin {part_tuples}")
-            dim = Dimension(actual_value_x[0], actual_value_y[0], actual_value_x[1])
-        return dim
-
-    @classmethod
-    def from_str(cls, text: str):
-        text = text.casefold()
-        parts = text.split("x")
-        if len(parts) == 0:
-            raise ValueError(f"Invalid margin string {text}")
-        if len(parts) > 2:
-            raise ValueError(f"Margin can have up to 2 parts ({text})")
-
-        part_tuples = [get_value_unit(x) for x in parts]
-        horizontal_tuple = part_tuples[0]
-        vertical_tuple = part_tuples[1] if len(part_tuples) > 1 else None
-        dim = cls.from_tuples(horizontal_tuple, vertical_tuple)
-        return dim
-
-    def __repr__(self):
-        return self.to_unit_str(None)
-
-    def to_unit_str(self, unit: str | None = None):
-        values = (self.horizontal, self.vertical)
-        if unit:
-            unit = unit.casefold().strip()
-            if unit == "cm":
-                values = [x / PT_BY_CM for x in values]
-            elif unit == "mm":
-                values = [x / PT_BY_MM for x in values]
-            elif unit == "inch" or unit == '"':
-                values = [x / PT_BY_INCH for x in values]
-            elif unit == "pt" or unit == "":
-                unit = "pt"
-                pass
-            else:
-                raise ValueError(f"Invalid unit type {unit}")
-        unit = unit or "pt"
-        return f"{values[0]}{unit} x {values[1]}{unit}"
+from .dimension import Dimension
 
 
 def expand_path(path: str | Path):
@@ -95,6 +15,7 @@ def expand_path(path: str | Path):
 @dataclass
 class Config:
     _output_folder: str = "~"
+    _libreoffice_path: str = r"%PROGRAMFILES(X86)%\LibreOffice\program\soffice.exe"
 
     _margin: str | Dimension = "0x0"
     _page_size: str | Dimension = "A4"
@@ -106,6 +27,18 @@ class Config:
         if isinstance(self._margin, str):
             return Dimension.from_str(self._margin)
         raise ValueError()
+
+    @property
+    def libreoffice_path(self):
+        p = Path(expand_path(self._libreoffice_path))
+        if not p.exists:
+            return None
+        return p.absolute()
+
+    @libreoffice_path.setter
+    def libreoffice_path(self, path: str | Path | None):
+        if path:
+            self._libreoffice_path = str(path)
 
     @property
     def page_size(self):
@@ -147,8 +80,11 @@ class Config:
         d = document()
         d.add(comment("Configuration file for stitcher"))
         d.add(nl())
+        d.add(comment("Path to Libre Office executable"))
+        d.add(comment("If the path is not valid or does not exists, conversion of LibreOffice formats is disabled."))
+        d.add("libreoffice_path", item(self._libreoffice_path))
         d.add(comment("Path to the output folder. May contain '~' and $ variables."))
-        d.add("output_folder", item(str(self._output_folder)))
+        d.add("output_folder", item(self._output_folder))
         d.add(nl())
         d.add(comment("Margin to be used when adding images. Horizontal then vertical dimension, separated by 'x'"))
         d.add("margin", item(str(self.margin)))
@@ -165,3 +101,4 @@ class Config:
             self.margin = doc.get("margin")
             self.page_size = doc.get("page_size")
             self.output_folder = doc.get("output_folder")
+            self.libreoffice_path = doc.get("libreoffice_path")
